# docs/estimator-openapi.yaml
openapi: 3.1.0
info:
  title: ZelfWonen deterministic estimator API
  version: 1.0.0
  description: Internal REST contract between Next.js and the Python regression service.
servers:
  - url: http://estimator:8000
paths:
  /v1/health:
    get:
      operationId: health
      responses:
        "200":
          description: Service and model are ready
          content:
            application/json:
              schema:
                type: object
                required: [status, modelVersion]
                properties:
                  status: { const: ready }
                  modelVersion: { type: string }
  /v1/estimate:
    post:
      operationId: estimateProperty
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateRequest"
      responses:
        "200":
          description: Deterministic numerical estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateResponse"
        "422":
          description: Pydantic validation error
        "503":
          description: Model unavailable
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    QualitativeFeatures:
      type: object
      additionalProperties: false
      required:
        - kitchenCondition
        - bathroomCondition
        - interiorFinish
        - naturalLight
        - exteriorCondition
        - confidence
        - evidence
      properties:
        kitchenCondition: { type: integer, minimum: 1, maximum: 5 }
        bathroomCondition: { type: integer, minimum: 1, maximum: 5 }
        interiorFinish: { type: integer, minimum: 1, maximum: 5 }
        naturalLight: { type: integer, minimum: 1, maximum: 5 }
        exteriorCondition: { type: integer, minimum: 1, maximum: 5 }
        confidence: { type: number, minimum: 0, maximum: 1 }
        evidence:
          type: array
          maxItems: 10
          items: { type: string, maxLength: 240 }
    EstimateRequest:
      type: object
      additionalProperties: false
      required:
        - postcode
        - houseNumber
        - propertyType
        - livingAreaSqm
        - roomCount
        - qualitativeFeatures
      properties:
        postcode:
          type: string
          pattern: "^[1-9][0-9]{3}[A-Z]{2}$"
        houseNumber: { type: integer, minimum: 1 }
        propertyType:
          type: string
          enum: [HOUSE, APARTMENT, PARKING, LAND, COMMERCIAL, OTHER]
        livingAreaSqm: { type: number, exclusiveMinimum: 0, maximum: 10000 }
        roomCount: { type: integer, minimum: 1, maximum: 100 }
        constructionYear: { type: integer, minimum: 1000, maximum: 2200 }
        qualitativeFeatures:
          oneOf:
            - $ref: "#/components/schemas/QualitativeFeatures"
            - type: "null"
    EstimateResponse:
      type: object
      additionalProperties: false
      required:
        - estimatedValueCents
        - lowerBoundCents
        - upperBoundCents
        - confidence
        - modelVersion
      properties:
        estimatedValueCents: { type: integer, minimum: 0 }
        lowerBoundCents: { type: integer, minimum: 0 }
        upperBoundCents: { type: integer, minimum: 0 }
        confidence: { type: number, minimum: 0, maximum: 1 }
        modelVersion: { type: string, minLength: 1 }